[multi-lora] operations: split the round executor from the queue-owning backend
The backend answers what is runnable (queues, rotation, collect,
write-back, depths, lifecycle fencing); the executor answers how one round
runs (payload parsing, sample conversion, padding, trainer calls, control
RPC routing, per-op outcomes) - the same buffer/consumer layering the
fully-async RL path uses between fully_async_data_buffer and its training
step. Pure move, plus the stale placeholder expectation in the driver
branch test now asserts the awaitable entry point instead.
